Improve validation (#2)
* dump operators in validation script


* dump subscriptions in validation script


* lengthened timeout for more validation testing


* changed structure of values.yaml and related templates b/c of bad output from the yamlencode function


* wrapped VALUES_CONTENT in quotes


* removed namespace cr


* updated validation script to detect successful cpd-operator subscription


* removed 10m sleep from debugging


* added debug output


* removed default values.yaml


* namespace var


* added namespace var default value

+++ .github/scripts/validate-deploy.sh
@@ -51,21 +51,20 @@ else
   sleep 30
 fi
 
-# DEPLOYMENT="${COMPONENT_NAME}-${BRANCH}"
-# count=0
-# until kubectl get deployment "${DEPLOYMENT}" -n "${NAMESPACE}" || [[ $count -eq 20 ]]; do
-#   echo "Waiting for deployment/${DEPLOYMENT} in ${NAMESPACE}"
-#   count=$((count + 1))
-#   sleep 15
-# done
 
-# if [[ $count -eq 20 ]]; then
-#   echo "Timed out waiting for deployment/${DEPLOYMENT} in ${NAMESPACE}"
-#   kubectl get all -n "${NAMESPACE}"
-#   exit 1
-# fi
 
-# kubectl rollout status "deployment/${DEPLOYMENT}" -n "${NAMESPACE}" || exit 1
+count=0
+until kubectl get subscription "cpd-operator" -n "${NAMESPACE}" || [[ $count -eq 20 ]]; do
+  echo "Waiting for subscription/cpd-operator in ${NAMESPACE}"
+  count=$((count + 1))
+  sleep 15
+done
+
+if [[ $count -eq 20 ]]; then
+  echo "Timed out waiting for subscription/cpd-operator in ${NAMESPACE}"
+  kubectl get all -n "${NAMESPACE}"
+  exit 1
+fi
 
 cd ..
 rm -rf .testrepo
